requirements.txt
# ── Core ML ──────────────────────────────────────────────────────────────────
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
optuna>=3.4.0

# ── Data Processing ───────────────────────────────────────────────────────────
pandas>=2.0.0
numpy>=1.24.0

# ── Streamlit UI ──────────────────────────────────────────────────────────────
streamlit>=1.32.0

# ── Visualisation ─────────────────────────────────────────────────────────────
plotly>=5.18.0
matplotlib>=3.7.0
seaborn>=0.12.0

# ── Explainability ────────────────────────────────────────────────────────────
shap>=0.44.0

# ── Model Persistence ─────────────────────────────────────────────────────────
joblib>=1.3.0

# ── Configuration ─────────────────────────────────────────────────────────────
PyYAML>=6.0

# ── Optional: GPU detection via PyTorch ──────────────────────────────────────
# Uncomment the line below if you want torch-based GPU detection.
# torch>=2.0.0

# ── Backend API ───────────────────────────────────────────────────────────────
fastapi>=0.116.0
uvicorn>=0.35.0
python-multipart>=0.0.20
